+++ main/packages/reactivizer/__tests__/forkjoin.test.ts
@@ -43,15 +43,23 @@ describe('forkjoin worker', () => {
     await forkMergeSort('singleWorker');
   }, 50000);
 
-  it('Exclude main thread', async () => {
+  it.skip('Exclude main thread', async () => {
     await forkMergeSort('excludeMainThread');
   }, 50000);
 
 
-  it('Scheduled workers can fork another worker or main worker itself', async () => {
+  it.skip('Scheduled workers can fork another worker or main worker itself', async () => {
     await forkMergeSort('scheduler');
   }, 40000);
 
+  it('Exclude main thread', async () => {
+    await forkMergeSort('excludeMainThread', undefined, 2000);
+  }, 50000);
+
+
+  it('Scheduled workers can fork another worker or main worker itself', async () => {
+    await forkMergeSort('scheduler', undefined, 2000);
+  }, 40000);
 });