+++ main/packages/reactivizer/es/fork-join/node-worker-broker.js
@@ -30,9 +30,19 @@ export function createBroker(workerController, opts) {
         const chan = new MessageChannel();
         portOfWorker.set(worker, chan.port1);
         o.dp.portOfWorker(portOfWorker);
-        const wo = new RxController({ name: '#' + workerNo + ' worker output', debug: opts === null || opts === void 0 ? void 0 : opts.debug, log: opts === null || opts === void 0 ? void 0 : opts.log });
+        const wo = new RxController({
+            name: '#' + workerNo + ' worker output',
+            debug: opts === null || opts === void 0 ? void 0 : opts.debug,
+            log: opts === null || opts === void 0 ? void 0 : opts.log,
+            debugExcludeTypes: opts === null || opts === void 0 ? void 0 : opts.debugExcludeTypes
+        });
         workerOutputs.set(workerNo, wo);
-        const wi = new RxController({ name: '#' + workerNo + ' worker input', debug: opts === null || opts === void 0 ? void 0 : opts.debug, log: opts === null || opts === void 0 ? void 0 : opts.log });
+        const wi = new RxController({
+            name: '#' + workerNo + ' worker input',
+            debug: opts === null || opts === void 0 ? void 0 : opts.debug,
+            log: opts === null || opts === void 0 ? void 0 : opts.log,
+            debugExcludeTypes: opts === null || opts === void 0 ? void 0 : opts.debugExcludeTypes
+        });
         wiByWorkerNo.set(workerNo, wi);
         o.dp.workerInputs(wiByWorkerNo);
         chan.port1.on('message', (event) => {