+++ main/packages/reactivizer/src/stream-core.ts
@@ -74,7 +74,7 @@ export class ControllerCore<I extends ActionFunctions = {[k: string]: never}> {
           rx.tap(action => {
             const type = nameOfAction(action);
             if (!this.debugExcludeSet.has(type)) {
-              opts.log!(this.logPrefix + 'rx:action', type, actionMetaToStr(action), ...(opts.logStyle === 'noParam' ? [] : action.p));
+              opts.log!(this.logPrefix, 'rx:action', type, actionMetaToStr(action), ...(opts.logStyle === 'noParam' ? [] : action.p));
             }
           }) :
           (typeof window !== 'undefined') || (typeof Worker !== 'undefined') ?
@@ -82,7 +82,7 @@ export class ControllerCore<I extends ActionFunctions = {[k: string]: never}> {
               const type = nameOfAction(action);
               if (!this.debugExcludeSet.has(type)) {
                 // eslint-disable-next-line no-console
-                console.log(`%c ${this.logPrefix}rx:action`, 'color: black; background: #8c61ff;',
+                console.log(`%c ${this.logPrefix} rx:action`, 'color: black; background: #8c61ff;',
                   type, actionMetaToStr(action), ...(opts.logStyle === 'noParam' ? [] : action.p));
               }
             }) :
@@ -90,7 +90,7 @@ export class ControllerCore<I extends ActionFunctions = {[k: string]: never}> {
               const type = nameOfAction(action);
               if (!this.debugExcludeSet.has(type)) {
                 // eslint-disable-next-line no-console
-                console.log( this.logPrefix + 'rx:action', type, actionMetaToStr(action), ...(opts.logStyle === 'noParam' ? [] : action.p));
+                console.log(this.logPrefix, 'rx:action', type, actionMetaToStr(action), ...(opts.logStyle === 'noParam' ? [] : action.p));
               }
             })
       )